@@ -14,7 +14,7 @@ import android.content.Context
 import android.os.Build
 import android.os.Environment
 import android.os.StatFs
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import com.example.util.Fields
 import com.example.util.logE
 import java.io. *
