@@ -14,7 +14,7 @@ import android.app.Activity
 import android.app.AlertDialog
 import android.content.Context
 import android.os.Looper
-import android.support.annotation.StringRes
+import androidx.annotation.StringRes
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
